refactor: move separate conent component to column component
