+++ admin/src/common/blocks/ColumnsBlock.tsx
@@ -1,5 +1,15 @@
-import { ColumnsLayoutPreview, ColumnsLayoutPreviewContent, ColumnsLayoutPreviewSpacing, createColumnsBlock } from "@comet/blocks-admin";
-import { ColumnsContentBlock } from "@src/common/blocks/ColumnsContentBlock";
+import {
+    ColumnsLayoutPreview,
+    ColumnsLayoutPreviewContent,
+    ColumnsLayoutPreviewSpacing,
+    createBlocksBlock,
+    createColumnsBlock,
+} from "@comet/blocks-admin";
+import { AnchorBlock } from "@comet/cms-admin";
+import { CallToActionListBlock } from "@src/common/blocks/CallToActionListBlock";
+import { HeadingBlock } from "@src/common/blocks/HeadingBlock";
+import { RichTextBlock } from "@src/common/blocks/RichTextBlock";
+import { SpaceBlock } from "@src/common/blocks/SpaceBlock";
 import { FormattedMessage } from "react-intl";
 
 const oneColumnLayouts = [
@@ -97,6 +107,17 @@ const twoColumnLayouts = [
         },
     },
 ];
+
+const ColumnsContentBlock = createBlocksBlock({
+    name: "ColumnsContent",
+    supportedBlocks: {
+        anchor: AnchorBlock,
+        richtext: RichTextBlock,
+        space: SpaceBlock,
+        heading: HeadingBlock,
+        callToActionList: CallToActionListBlock,
+    },
+});
 
 export const ColumnsBlock = createColumnsBlock({
     name: "Columns",
